Fixed a major issue, added new function, small improvements
Fixed an issue introduced accidentally where OBBfilePath was being checked to have nothing before "com.", when it should be OBBname being checked for that condition.
Added installAgain function to give the user the option install the same build on another device without restarting the whole process.
Removed countdown after fatal errors and made the error code a little easier to see for end-users.